Break X11 transient windows to remove #include <X11/Xlib.h> in API.
`Xlib.h` has typedefs using the same names on ClanLib (like `clan::Font` and
X11's `Font`) which causes ambiguity problems when the `clan` namespace is
imported into the global namespace.

There is no API to generically move WM handles, but it seems that it is only
needed when creating multiple WM windows linked to the same top-level parent
window on X11, thus we are disabling it as a temporary workaround.

The Windows build should not affected by this change.