Added DisplayWindowHandle + implementation for X11
This class is used to contain the display manager's window
handle so that ClanLib can obtain it from a `DisplayWindow`
instance. It is implemented as a pointer to an undefined
class declaration on the API which is then defined inside
the whatever windowing system that is compiled with the
library.