Re-implement pointing device coordinates for InputDevice and InputEvent.
The unit used for mouse position has switched from Device Dependent Pixel
(hardware pixel unit) to Device Independent Pixel (DPI-scaled pixel unit) in
the DPR patch on January. This commit brings back DDP positioning for pointing
devices with the addition of `InputDevice::get_device_position()` and
`InputEvent::mouse_device_pos`. The UI module is not affected by this change.

Also in this commit are some cleanup on the affected files.