Prevent D3D to process alt+enter messages
D3D target has an unpleasant feature by default - it seeks for alt+enter
sequence and process it before these messages got dispatched to the
actual program. It comes to a nasty effect of having alt+enter
fullscreen mode without programmer knowing about it.
This is a fix for it.