Fixed bug The old scheme - draw all backgrounds, all borders and all contents. The new scheme - draw the background, border and contents root View, then background, borders and contents of the first child and so on - like in browsers.
The bug can see in HelloWorld - on the scrollbar there is no grip in the old sheme, because for grip the image sets as background-image, but for track and thumb as border-image.

Also speedup the ScrollBarView - now position is changed without layout and redrawing the whole window.

And fix small bug in ScrollBarView - crash on set_disabled(). Next time will be fix the more serious bugs - wrong positioning and timer's non-common behaviour.